Run DNS resolution locally
This allows implementing it via InetAddres/getaddr, which
can be configured.

+++ library/src/commonMain/kotlin/com/tidal/networktime/internal/NtpExchanger.kt
@@ -10,7 +10,7 @@ internal class NtpExchanger(
   operator fun invoke(
     address: String,
     queryTimeout: Duration,
-    ntpVersion: UInt,
+    ntpVersion: UByte,
   ): NtpExchangeResult? {
     val ntpUdpSocketOperations = NtpUdpSocketOperations()
     val requestPacket = NtpPacket(
